requirements.txt
aioamqp==0.14.0
aiohttp==3.7.3
click==7.1.2
pyyaml==5.3.1
requests==2.25.1